[PS] Rename AbstractCachingDataSource to CachedRepositoryDataSource to reflect it's purpose and made it a trait. MemoryCached and FileCached are new names for data sources that implement this trait
